feat: publish storage and document delivery manifests
