+++ packages/public-document-delivery/tests/unit/public-document-delivery.test.ts
@@ -11,7 +11,7 @@ import {
   type RevokePublicDocumentDeliveryGrantInput,
   resolvePublicDocumentDeliveryGrant,
   revokePublicDocumentDeliveryGrant,
-} from "../../src/public-document-delivery.js"
+} from "../../src/index.js"
 
 function createMemoryGrantStore(): PublicDocumentDeliveryGrantStore & {
   grants: PublicDocumentDeliveryGrant[]
